+++ src/main/java/org/anch/dao/CustomerDao.java
@@ -4,6 +4,6 @@ import org.anch.entity.Customer;
 
 import java.util.Optional;
 
-public interface CustomerDao extends CrudDao<Short, Customer> {
+public interface CustomerDao extends CrudDao<Integer, Customer> {
 
 }
