+++ src/main/java/org/anch/entity/Customer.java
@@ -20,7 +20,7 @@ public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "customer_id")
-    private Short id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "store_id", nullable = false)
